requirements.txt
# Goldbach Unified Trading Platform
# Requirements

# Web Framework
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-socketio>=5.8.0
eventlet>=0.33.0
gunicorn>=21.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Data Sources
yfinance>=0.2.28
requests>=2.31.0

# Database
# SQLite is built-in

# Environment
python-dotenv>=1.0.0

# Charts (optional, for server-side rendering)
matplotlib>=3.7.0

# Testing
pytest>=7.4.0

# Utilities
python-dateutil>=2.8.2

# Production
gevent>=23.0.0
gevent-websocket>=0.10.1
